+++ app/opulence/http/requests/UploadException.php
@@ -1,0 +1,13 @@
+<?php
+/**
+ * Copyright (C) 2015 David Young
+ *
+ *
+ */
+namespace Opulence\HTTP\Requests;
+use Exception;
+
+class UploadException extends Exception
+{
+    // Don't do anything
+}