+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+<h2>v0.6.6</h2>
+
+<h3>HTTP</h3>
+* Added `Opulence\HTTP\Requests\UploadedFile` and `FileParameters`
+* Uploaded files now instantiate `UploadedFile` objects rather than simple arrays
+
 <h2>v0.6.5</h2>
 
 <h3>Environment</h3>